+++ scripts/test.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 
-
-# @todo: add .ape test cases
-
-cargo test
+cargo run run tests/if.ape
+cargo run run tests/while.ape
+cargo run run tests/loop.ape
+cargo run run tests/variables.ape
+cargo run run tests/functions.ape
+cargo run run tests/mod.ape